Moved new lib tests to tests/lib.rs removed boilerplate code
