+++ pallets/subtensor/src/lib.rs
@@ -493,7 +493,7 @@ impl<T: Trait> Module<T> {
 }
 
 #[derive(Encode, Decode, Clone, Eq, PartialEq)]
-pub struct FeeFromSelfEmission<T: Trait + Send + Sync>(PhantomData<T>);
+pub struct FeeFromSelfEmission<T: Trait + Send + Sync>(pub PhantomData<T>);
 
 impl<T: Trait + Send + Sync> sp_std::fmt::Debug for FeeFromSelfEmission<T> {
 	fn fmt(&self, f: &mut sp_std::fmt::Formatter) -> sp_std::fmt::Result {
@@ -533,223 +533,5 @@ where
 			}
 			_ => Ok(Default::default()),
 		}
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	use frame_support::{
-		assert_ok, impl_outer_origin, parameter_types, impl_outer_dispatch,
-		weights::{DispatchInfo}, traits::{OnInitialize, OnFinalize}
-	};
-	use sp_core::H256;
-
-	// The testing primitives are very useful for avoiding having to work with signatures
-	// or public keys. `u64` is used as the `AccountId` and no `Signature`s are required.
-	use sp_runtime::{
-		Perbill,
-		testing::Header,
-		traits::{BlakeTwo256, IdentityLookup},
-	};
-
-	impl_outer_origin!{
-		pub enum Origin for Test where system = frame_system {}
-	}
-
-	impl_outer_dispatch! {
-		pub enum OuterCall for Test where origin: Origin {
-			self::SubtensorModule,
-		}
-	}
-
-	pub type Balances = pallet_balances::Module<Test>;
-
-	// For testing the pallet, we construct most of a mock runtime. This means
-	// first constructing a configuration type (`Test`) which `impl`s each of the
-	// configuration traits of pallets we want to use.
-	#[derive(Clone, Eq, PartialEq)]
-	pub struct Test;
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub const MaximumBlockWeight: Weight = 1024;
-		pub const MaximumBlockLength: u32 = 2 * 1024;
-		pub const AvailableBlockRatio: Perbill = Perbill::one();
-	}
-	impl frame_system::Trait for Test {
-		type BaseCallFilter = ();
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Call = OuterCall;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = ();
-		type BlockHashCount = BlockHashCount;
-		type MaximumBlockWeight = MaximumBlockWeight;
-		type DbWeight = ();
-		type BlockExecutionWeight = ();
-		type ExtrinsicBaseWeight = ();
-		type MaximumExtrinsicWeight = MaximumBlockWeight;
-		type MaximumBlockLength = MaximumBlockLength;
-		type AvailableBlockRatio = AvailableBlockRatio;
-		type Version = ();
-		type PalletInfo = ();
-		type AccountData = pallet_balances::AccountData<u64>;
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-	}
-	parameter_types! {
-		pub const ExistentialDeposit: u64 = 1;
-	}
-	impl pallet_balances::Trait for Test {
-		type MaxLocks = ();
-		type Balance = u64;
-		type DustRemoval = ();
-		type Event = ();
-		type ExistentialDeposit = ExistentialDeposit;
-		type AccountStore = System;
-		type WeightInfo = ();
-	}
-	impl Trait for Test {
-		type Event = ();
-		type Currency = Balances;
-	}
-	type System = frame_system::Module<Test>;
-	type SubtensorModule = Module<Test>;
-
-	#[allow(dead_code)]
-	pub fn subscribe_neuron(hotkey_account_id : u64, ip: u128, port: u16, ip_type : u8, modality: u8, coldkey_acount_id : u64) -> NeuronMetadata<u64> {
-		let result = SubtensorModule::subscribe(<<Test as system::Trait>::Origin>::signed(hotkey_account_id), ip, port, ip_type, modality, coldkey_acount_id);
-		assert_ok!(result);
-		let neuron = SubtensorModule::get_neuron_for_hotkey(&hotkey_account_id);
-		neuron
-	}
-
-	// Build genesis storage according to the mock runtime.
-	pub fn new_test_ext() -> sp_io::TestExternalities {
-		system::GenesisConfig::default().build_storage::<Test>().unwrap().into()
-	}
-
-	#[allow(dead_code)]
-	pub fn run_to_block(n: u64) {
-		while System::block_number() < n {
-			SubtensorModule::on_finalize(System::block_number());
-			System::on_finalize(System::block_number());
-			System::set_block_number(System::block_number() + 1);
-			System::on_initialize(System::block_number());
-			SubtensorModule::on_initialize(System::block_number());
-		}
-	}
-
-	#[test]
-	fn fee_from_emission_works() {
-		new_test_ext().execute_with(|| {
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-			assert!( FeeFromSelfEmission::<Test>(PhantomData).validate(&1, &call, &info, len).is_ok() );
-		});
-	}
-
-	#[test]
-	fn fee_from_emission_priority_no_neuron() {
-		new_test_ext().execute_with(|| {
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&1, &call, &info, len).unwrap().priority, 0);
-		});
-	}
-
-	#[test]
-	fn fee_from_emission_priority_with_neuron() {
-		new_test_ext().execute_with(|| {
-
-			let hotkey_account_id = 1;
-			subscribe_neuron(hotkey_account_id, 10, 666, 4, 0, 66);
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&hotkey_account_id, &call, &info, len).unwrap().priority, 0);
-		});
-	}
-
-
-	#[test]
-	fn fee_from_emission_priority_with_neuron_and_weights() {
-		new_test_ext().execute_with(|| {
-
-			let hotkey_account_id = 1;
-			let neuron = subscribe_neuron(hotkey_account_id, 10, 666, 4, 0, 66);
-
-			let weight_uids = vec![neuron.uid];
-			let weight_values = vec![u32::MAX];
-			assert_ok!(SubtensorModule::set_weights(<<Test as system::Trait>::Origin>::signed(hotkey_account_id), weight_uids.clone(), weight_values.clone()));
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&hotkey_account_id, &call, &info, len).unwrap().priority, 0);
-		});
-	}
-
-	#[test]
-	fn fee_from_emission_priority_with_neuron_and_weights_and_stake() {
-		new_test_ext().execute_with(|| {
-
-			let hotkey_account_id = 1;
-			let neuron = subscribe_neuron(hotkey_account_id, 10, 666, 4, 0, 66);
-
-			let weight_uids = vec![neuron.uid];
-			let weight_values = vec![u32::MAX];
-			assert_ok!(SubtensorModule::set_weights(<<Test as system::Trait>::Origin>::signed(hotkey_account_id), weight_uids.clone(), weight_values.clone()));
-
-
-			SubtensorModule::add_stake_to_neuron_hotkey_account(neuron.uid, 100000000); // Add the stake.
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&hotkey_account_id, &call, &info, len).unwrap().priority, 0);
-		});
-	}
-
-	#[test]
-	fn fee_from_emission_priority_with_neuron_and_weights_and_stake_and_run_to_block() {
-		new_test_ext().execute_with(|| {
-
-			let hotkey_account_id = 1;
-			let neuron = subscribe_neuron(hotkey_account_id, 10, 666, 4, 0, 66);
-
-			let weight_uids = vec![neuron.uid];
-			let weight_values = vec![u32::MAX];
-			assert_ok!(SubtensorModule::set_weights(<<Test as system::Trait>::Origin>::signed(hotkey_account_id), weight_uids.clone(), weight_values.clone()));
-
-
-			SubtensorModule::add_stake_to_neuron_hotkey_account(neuron.uid, 1000000000); // Add the stake.
-
-			let call = <Call<Test>>::set_weights(vec![0], vec![0]).into();
-			let info = DispatchInfo::default();
-			let len = 10;
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&hotkey_account_id, &call, &info, len).unwrap().priority, 0);
-
-			run_to_block(1);
-
-			assert_eq!( FeeFromSelfEmission::<Test>(PhantomData).validate(&hotkey_account_id, &call, &info, len).unwrap().priority, 50000000);
-		});
 	}
 }
